+++ tests/test_main.py
@@ -1,6 +1,13 @@
 import pytest
-from module.main import main
+from module.product import Product
+from module.category import Category
 
+product1 = Product("Samsung Galaxy S23 Ultra", "256GB, Серый цвет, 200MP камера", 180000.0, 5)
+product2 = Product("Iphone 15", "512GB, Gray space", 210000.0, 8)
+product3 = Product("Xiaomi Redmi Note 11", "1024GB, Синий", 31000.0, 14)
+category1 = Category("Смартфоны", "Смартфоны, как средство не только коммуникации, но и получения дополнительных функций для удобства жизни", [product1, product2, product3])
+product4 = Product("55 QLED 4K", "Фоновая подсветка", 123000.0, 7)
+category2 = Category("Телевизоры", "Современный телевизор, который позволяет наслаждаться просмотром, станет вашим другом и помощником", [product4])
 def test_main(product1, product2, product3, product4, category1, category2):
     assert product1.name == "Samsung Galaxy S23 Ultra"
     assert product1.description == "256GB, Серый цвет, 200MP камера"
